Major rearrangement of classes
