+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:theme="@style/AppTheme"
         android:usesCleartextTraffic="true">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:configChanges="keyboardHidden|orientation|screenSize"
             android:icon="@mipmap/ic_launcher"
             android:label="@string/app_name"
@@ -31,50 +31,50 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".AboutActivity"
+            android:name=".ui.AboutActivity"
             android:icon="@drawable/appicon"
             android:label="@string/about_title"
-            android:parentActivityName=".MainActivity"
+            android:parentActivityName=".ui.MainActivity"
             android:theme="@style/AppTheme">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.workingagenda.democracydroid.MainActivity" />
+                android:value="com.workingagenda.democracydroid.ui.MainActivity" />
         </activity>
         <activity
-            android:name=".StoryActivity"
+            android:name=".ui.StoryActivity"
             android:icon="@drawable/appicon"
             android:label="@string/app_name"
-            android:parentActivityName=".MainActivity"
+            android:parentActivityName=".ui.MainActivity"
             android:theme="@style/AppTheme">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.workingagenda.democracydroid.MainActivity" />
+                android:value="com.workingagenda.democracydroid.ui.MainActivity" />
         </activity>
         <activity
-            android:name=".MediaActivity"
+            android:name=".ui.player.MediaActivity"
             android:configChanges="keyboardHidden|orientation|screenSize"
             android:icon="@drawable/appicon"
             android:label="@string/app_name"
             android:launchMode="singleTask"
-            android:parentActivityName=".MainActivity"
+            android:parentActivityName=".ui.MainActivity"
             android:theme="@style/AppTheme">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.workingagenda.democracydroid.MainActivity" />
+                android:value="com.workingagenda.democracydroid.ui.MainActivity" />
         </activity>
 
         <activity
-            android:name=".SettingsActivity"
+            android:name=".ui.settings.SettingsActivity"
             android:icon="@drawable/appicon"
             android:label="@string/name_setting"
-            android:parentActivityName=".MainActivity"
+            android:parentActivityName=".ui.MainActivity"
             android:theme="@style/AppTheme">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.workingagenda.democracydroid.MainActivity" />
+                android:value="com.workingagenda.democracydroid.ui.MainActivity" />
         </activity>
 
-        <service android:name=".MediaService">
+        <service android:name=".core.service.MediaService">
             <intent-filter>
                 <action android:name=".MediaService" />
             </intent-filter>
