--- app/src/main/java/com/workingagenda/democracydroid/Adapters/GridSpacingItemDecoration.java
@@ -1,38 +1,0 @@
-package com.workingagenda.democracydroid.Adapters;
-
-import android.graphics.Rect;
-import android.view.View;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-/**
- * Created by derrickrocha on 7/16/17.
- */
-public class GridSpacingItemDecoration extends RecyclerView.ItemDecoration {
-
-    private final int spanCount;
-    private final int spacing;
-    private final boolean includeEdge;
-
-    public GridSpacingItemDecoration(int spanCount, int spacing, boolean includeEdge) {
-        this.spanCount = spanCount;
-        this.spacing = spacing;
-        this.includeEdge = includeEdge;
-    }
-
-    @Override
-    public void getItemOffsets(@NonNull Rect outRect, @NonNull View view, RecyclerView parent,
-                               @NonNull RecyclerView.State state) {
-        int position = parent.getChildAdapterPosition(view); // item position
-        int column = position % spanCount; // item column
-
-        if (includeEdge) {
-            outRect.left = spacing - column * spacing / spanCount; // spacing - column * ((1f / spanCount) * spacing)
-            outRect.right = (column + 1) * spacing / spanCount; // (column + 1) * ((1f / spanCount) * spacing)
-        } else {
-            outRect.left = column * spacing / spanCount; // column * ((1f / spanCount) * spacing)
-            outRect.right = spacing - (column + 1) * spacing / spanCount; // spacing - (column + 1) * ((1f /    spanCount) * spacing)
-        }
-    }
-}
