--- app/src/main/java/com/workingagenda/democracydroid/SettingsActivity.java
@@ -1,40 +1,0 @@
-package com.workingagenda.democracydroid;
-
-import android.os.Bundle;
-import android.view.MenuItem;
-
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.NavUtils;
-
-import com.workingagenda.democracydroid.databinding.ActivitySettingsBinding;
-
-/**
- * Created by fen on 5/15/16.
- */
-public class SettingsActivity extends AppCompatActivity {
-    ActivitySettingsBinding binding;
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        binding = ActivitySettingsBinding.inflate(getLayoutInflater());
-        setContentView(binding.getRoot());
-
-        setSupportActionBar(binding.settingsToolbar);
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-
-        getSupportFragmentManager()
-                .beginTransaction()
-                .replace(R.id.settings_container, new SettingsFragment())
-                .commit();
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        if (item.getItemId() == android.R.id.home) {
-            NavUtils.navigateUpFromSameTask(this);
-            return true;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-}
