+++ app/src/main/java/com/workingagenda/democracydroid/util/Utils.java
@@ -1,0 +1,14 @@
+package com.workingagenda.democracydroid.util;
+
+import android.util.DisplayMetrics;
+import android.util.TypedValue;
+
+public class Utils {
+
+    /**
+     * Converting dp to pixel
+     */
+    public static int dpToPx(int dp, DisplayMetrics metrics) {
+        return Math.round(TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, metrics));
+    }
+}
