+++ app/src/main/res/layout/activity_about.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.workingagenda.democracydroid.AboutActivity">
+    tools:context="com.workingagenda.democracydroid.ui.AboutActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
