+++ app/src/main/res/layout/activity_story.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.workingagenda.democracydroid.StoryActivity">
+    tools:context="com.workingagenda.democracydroid.ui.StoryActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
